+++ test/libsolidity/syntaxTests/inheritance/override/override_multiple.sol
@@ -1,5 +1,4 @@
 abstract contract A {
-	int public testvar;
 	function foo() internal virtual returns (uint256);
 	function test(uint8 _a) internal virtual returns (uint256) {}
 }
@@ -18,4 +17,3 @@ contract X is A, B, C, D {
 	function test() internal override returns (uint256) {}
 	function foo() internal override(A, B, C, D) returns (uint256) {}
 }
-// ----