+++ test/libsolidity/syntaxTests/inheritance/override/override_public_vars.sol
@@ -1,7 +1,0 @@
-abstract contract A {
-	int public testvar;
-}
-abstract contract X is A {
-	int public override testvar;
-}
-// ----