+++ test/libsolidity/syntaxTests/inheritance/shadowing_private_base_state_vars.sol
@@ -1,0 +1,6 @@
+contract A {
+	uint private i;
+}
+contract B is A {
+	uint i;
+}